mata: Bringup telephony after U QPR2
Notably, pd-mapper service must be kept for cnss power voting.

This is a large, squashed change that includes everything needed to run
the new radio stack on mata. Apart from the original change by Bruno
Martins for pro1, the following changes are squashed in:
- I38329f6d4ebb7b1a0b0f01e40918fdb73ccbcba9 ("Move telephony
packages to /system_ext")
- I22c2ad6b6da0aae9385631f7f01a1eae7eb59d75 ("Drop no longer used QTI
telephony jar")

Change-Id: Id272a09338b285d419f8f012cfc6b62a7f8ef50c